+++ packages/gamut-patterns/src/patterns/grid/index.tsx
@@ -1,0 +1,3 @@
+export * from './GridDense';
+export * from './GridLoose';
+export * from './GridRegular';