+++ packages/gamut/src/Dialog/index.tsx
@@ -1,119 +1,70 @@
 import { MiniDeleteIcon } from '@codecademy/gamut-icons';
-import { pxRem, variant } from '@codecademy/gamut-styles';
-import { css } from '@emotion/react';
+import { system } from '@codecademy/gamut-styles';
+import { StyleProps } from '@codecademy/variance';
 import styled from '@emotion/styled';
 import React from 'react';
 
-import { Box, FlexBox } from '../Box';
 import { FillButton, IconButton, TextButton } from '../Button';
 import { Overlay, OverlayProps } from '../Overlay';
 import { Pattern } from '../Pattern';
-import { VisualTheme } from '../theming/VisualTheme';
+import { Text } from '../Typography';
 
-const modes = variant({
-  prop: 'mode',
+type SizeVariants = StyleProps<typeof modalSizeVariants>;
+
+const modalSizeVariants = system.variant({
+  prop: 'size',
   variants: {
-    light: {
-      textColor: 'navy',
-      backgroundColor: 'white',
-      borderColor: 'navy',
-    },
-    dark: {
-      textColor: 'white',
-      backgroundColor: 'navy',
-      borderColor: 'white',
-    },
+    small: { width: '400px', minHeight: '170px' },
+    medium: { width: '640px', minHeight: '240px' },
   },
 });
 
-/** These effects should be moved to theme */
-const shroudColor = {
-  dark: 'rgba(0,0,0, .75)',
-  light: 'rgba(255, 255, 255, 0.95)',
-};
+const ShroudedOverlay = styled(Overlay)(system.css({ bg: 'shadow-opaque' }));
 
-const ShroudedOverlay = styled(Overlay)<{ mode?: VisualTheme }>(
-  ({ mode = 'light' }) => `
-  background-color: ${shroudColor[mode]};
-`
+const ModalWrapper = styled.div<SizeVariants>(
+  system.css({
+    position: 'relative',
+    zIndex: 1,
+    maxWidth: 'calc(100vw - 2rem)',
+    borderRadius: '2px',
+    display: 'grid',
+  }),
+  modalSizeVariants
 );
-
-const ModalWrapper = styled.div(modes, ({ theme }) => {
-  return css`
-    position: relative;
-    z-index: 1;
-    max-width: calc(100vw - ${theme.spacing[32]});
-    width: ${pxRem(400)};
-    border-width: 1px;
-    border-style: solid;
-    border-radius: 2px;
-  `;
-});
 
 /** This will need to be consolidated with Card / CoachMark / Toast  */
 const ModalShadow = styled(Pattern)(
-  ({ theme }) => css`
-    width: 100%;
-    height: 100%;
-    position: absolute;
-    top: ${theme.spacing[8]};
-    left: -${theme.spacing[8]};
-  `
+  system.css({
+    textColor: 'text',
+    width: 1,
+    height: 1,
+    position: 'absolute',
+    top: '.5rem',
+    left: '-.5rem',
+  })
 );
 
-const ModalForeground = styled.div`
-  width: 100%;
-  height: 100%;
-  position: absolute;
-  top: 0;
-  left: 0;
-`;
-
-const ModalBody = styled.div(({ theme }) => {
-  return css`
-    position: relative;
-    background-color: inherit;
-    color: inherit;
-    display: grid;
-    padding: ${theme.spacing[24]};
-    grid-row-gap: ${theme.spacing[12]};
-    grid-column-gap: ${theme.spacing[16]};
-    grid-template-columns: 1fr min-content ${theme.spacing[32]};
-    grid-template-rows: repeat(3, auto);
-    grid-template-areas:
-      'title title close'
-      'children children children'
-      'cancel confirm confirm';
-  `;
-});
-
-const Title = styled(Box)`
-  grid-area: title;
-`;
-
-const CloseButton = styled.div`
-  grid-area: close;
-`.withComponent(IconButton);
-
-const Content = styled(Box)`
-  grid-area: children;
-`;
-
-const ConfirmButton = styled(FillButton)`
-  grid-area: confirm;
-`;
-
-const CancelButton = styled(TextButton)`
-  grid-area: cancel;
-`.withComponent((props: DialogProps['cancelCta']) => (
-  <FlexBox justifyContent="flex-end">
-    <TextButton {...props} />
-  </FlexBox>
-));
+const ModalBody = styled.div(
+  system.css({
+    border: 1,
+    position: 'relative',
+    bg: 'background',
+    textColor: 'text',
+    display: 'grid',
+    p: 24,
+    rowGap: 12,
+    columnGap: 16,
+    gridTemplateColumns: '1fr min-content 2rem',
+    gridTemplateRows: 'max-content 1fr max-content',
+    gridTemplateAreas: `'title title close'
+    'content content content'
+    'cancel confirm confirm'`,
+  })
+);
 
 export interface DialogProps
-  extends Pick<OverlayProps, 'clickOutsideCloses' | 'escapeCloses'> {
-  mode?: VisualTheme;
+  extends Pick<OverlayProps, 'clickOutsideCloses' | 'escapeCloses'>,
+    SizeVariants {
   isOpen: boolean;
   title: React.ReactNode;
   children: React.ReactNode;
@@ -129,12 +80,12 @@ export interface DialogProps
 }
 
 export const Dialog: React.FC<DialogProps> = ({
-  mode = 'light',
   title,
   children,
   confirmCta,
   cancelCta,
   onRequestClose,
+  size = 'small',
   ...rest
 }) => {
   const onConfirm = () => {
@@ -148,39 +99,38 @@ export const Dialog: React.FC<DialogProps> = ({
   };
 
   return (
-    <ShroudedOverlay mode={mode} onRequestClose={onCancel} {...rest}>
+    <ShroudedOverlay onRequestClose={onCancel} {...rest}>
       <ModalWrapper
-        mode={mode}
+        size={size}
         aria-hidden="false"
         aria-modal="true"
         role="dialog"
       >
         <ModalShadow name="checkerDense" />
-        <ModalForeground />
         <ModalBody>
-          <Title
-            as="h2"
-            margin={0}
-            fontSize={20}
-            lineHeight="base"
-            fontWeight="title"
-          >
+          <Text as="h2" fontSize={20} lineHeight="base" gridArea="title">
             {title}
-          </Title>
-          <CloseButton
+          </Text>
+          <IconButton
             aria-label="Close Dialog"
-            mode={mode}
             size="small"
+            variant="secondary"
             icon={MiniDeleteIcon}
             onClick={onCancel}
+            gridArea="close"
           />
-          <Content fontSize={16} lineHeight="base">
+          <Text as="div" gridArea="content" data-testid="dialog-content">
             {children}
-          </Content>
+          </Text>
           {cancelCta && (
-            <CancelButton mode={mode} {...cancelCta} onClick={onCancel} />
+            <TextButton
+              {...cancelCta}
+              onClick={onCancel}
+              justifySelf="end"
+              gridArea="cancel"
+            />
           )}
-          <ConfirmButton mode={mode} {...confirmCta} onClick={onConfirm} />
+          <FillButton {...confirmCta} onClick={onConfirm} gridArea="confirm" />
         </ModalBody>
       </ModalWrapper>
     </ShroudedOverlay>