+++ packages/styleguide/stories/Foundations/ColorMode/examples.tsx
@@ -1,13 +1,87 @@
 /* eslint-disable local-rules/gamut-import-paths */
-import { Anchor, Box, FlexBox, Text, Toggle } from '@codecademy/gamut/src';
+import {
+  Anchor,
+  Box,
+  CTAButton,
+  FillButton,
+  FlexBox,
+  GridBox,
+  IconButton,
+  Logo,
+  ProLabel,
+  StrokeButton,
+  Text,
+  TextButton,
+  Toggle,
+} from '@codecademy/gamut/src';
+import { MiniDeleteIcon, SearchIcon } from '@codecademy/gamut-icons';
 import { Background, ColorMode } from '@codecademy/gamut-styles/src';
 import React, { ComponentProps, useState } from 'react';
+
+const renderButtons = (variant?: 'primary' | 'secondary', disabled = false) => {
+  return (
+    <GridBox
+      rowGap={16}
+      columnGap={16}
+      gridTemplateColumns="repeat(5, max-content)"
+      gridAutoRows="3em"
+      alignItems="start"
+      justifyItems="start"
+      pt={16}
+    >
+      <GridBox gridRowEnd="span 2">
+        <CTAButton variant={variant as any} disabled={disabled}>
+          CTA
+        </CTAButton>
+      </GridBox>
+      <FillButton variant={variant} disabled={disabled}>
+        Fill
+      </FillButton>
+      <StrokeButton variant={variant} disabled={disabled}>
+        Stroke
+      </StrokeButton>
+      <TextButton variant={variant} disabled={disabled}>
+        Text
+      </TextButton>
+      <IconButton variant={variant} icon={SearchIcon} disabled={disabled} />
+      <FillButton variant={variant} size="small" disabled={disabled}>
+        Fill
+      </FillButton>
+      <StrokeButton variant={variant} size="small" disabled={disabled}>
+        Stroke
+      </StrokeButton>
+      <TextButton variant={variant} size="small" disabled={disabled}>
+        Text
+      </TextButton>
+      <IconButton
+        variant={variant}
+        size="small"
+        icon={MiniDeleteIcon}
+        disabled={disabled}
+      />
+    </GridBox>
+  );
+};
+
+const renderLinks = () => {
+  return (
+    <GridBox
+      mt={16}
+      columnGap={16}
+      gridTemplateColumns="repeat(3, max-content)"
+    >
+      <Anchor>Inline Link</Anchor>
+      <Anchor variant="standard">Standard Link</Anchor>
+      <Anchor variant="interface">Interface Link</Anchor>
+    </GridBox>
+  );
+};
 
 export const ColorModeExample = () => {
   const [isDark, setIsDark] = useState(false);
   return (
-    <Box marginY={16} marginBottom={32}>
-      <FlexBox fontWeight="title" marginBottom={8} alignItems="center">
+    <Box mt={16} mb={32}>
+      <FlexBox mb={8} alignItems="center">
         <Text mr={8} as="strong">
           Use Dark Mode
         </Text>
@@ -20,15 +94,21 @@ export const ColorModeExample = () => {
         />
       </FlexBox>
       <ColorMode mode={isDark ? 'dark' : 'light'}>
-        <Box
-          backgroundColor="background"
-          padding={24}
-          borderColor="text"
-          borderStyle="solid"
-          borderWidth="1px"
-        >
-          <Text as="h5" fontSize={26} fontWeight="title" mb={16}>
+        <Box bg="background" p={24} border={1}>
+          <Text as="h3" mb={24}>
             {isDark ? 'Dark' : 'Light'} Mode
+          </Text>
+          <GridBox
+            columnGap={16}
+            pb={16}
+            gridTemplateColumns="max-content max-content max-content"
+          >
+            <Logo variant="default" />
+            <Logo variant="pro" />
+            <Logo variant="mini" />
+          </GridBox>
+          <Text as="p" fontSize={16} fontFamily="accent" mb={16}>
+            <ProLabel height={22} verticalAlign="text-bottom" /> Cool Feature
           </Text>
           <Text as="p" mb={16}>
             Lorem ipsum dolor sit amet, sed do eiusmod tempor incididunt ut
@@ -36,6 +116,10 @@ export const ColorModeExample = () => {
             veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex
             ea commodo consequat.
           </Text>
+          {renderLinks()}
+          {renderButtons('primary')}
+          {renderButtons('secondary')}
+          {renderButtons('primary', true)}
         </Box>
       </ColorMode>
     </Box>
@@ -48,7 +132,7 @@ export const BackgroundExample: React.FC<ComponentProps<typeof Background>> = ({
 }) => {
   return (
     <Background {...rest} p={24}>
-      <Text as="p" fontSize={26} fontWeight="title" mb={16}>
+      <Text variant="title-md" mb={16}>
         {rest.bg}
       </Text>
       <Text as="p" mb={16}>