+++ packages/variance/CHANGELOG.md
@@ -3,6 +3,40 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.0](https://github.com/Codecademy/client-modules/compare/@codecademy/variance@0.11.2...@codecademy/variance@0.12.0) (2021-05-18)
+
+
+### Features
+
+* **theme:** createTheme self containment ([59ddff7](https://github.com/Codecademy/client-modules/commit/59ddff7224cd8e0972212dd6ba1696f40ccac835)), closes [#666666](https://github.com/Codecademy/client-modules/issues/666666) [#666666](https://github.com/Codecademy/client-modules/issues/666666)
+
+
+
+### [0.11.2](https://github.com/Codecademy/client-modules/compare/@codecademy/variance@0.11.1...@codecademy/variance@0.11.2) (2021-05-07)
+
+**Note:** Version bump only for package @codecademy/variance
+
+
+
+
+
+### [0.11.1](https://github.com/Codecademy/client-modules/compare/@codecademy/variance@0.11.0...@codecademy/variance@0.11.1) (2021-05-03)
+
+**Note:** Version bump only for package @codecademy/variance
+
+
+
+
+
+## [0.11.0](https://github.com/Codecademy/client-modules/compare/@codecademy/variance@0.10.1...@codecademy/variance@0.11.0) (2021-04-29)
+
+
+### Features
+
+* **Buttons:** ColorModes + Props + Refactor + Refs Oh My! [GM-211] [REACH-856] [GM-192] ([c4643db](https://github.com/Codecademy/client-modules/commit/c4643db64d3684319f8d5fc6f68d05197ec1f61e))
+
+
+
 ### [0.10.1](https://github.com/Codecademy/client-modules/compare/@codecademy/variance@0.10.0...@codecademy/variance@0.10.1) (2021-04-27)